Fixes in Welcome Page Activity
