+++ app/src/main/java/com/amciof/Yandex_MDS2019/SlidePagerAdapter.java
@@ -4,13 +4,11 @@ package com.amciof.Yandex_MDS2019;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentPagerAdapter;
-
 import java.util.List;
 
 
 
-public class SlidePagerAdapter
-        extends FragmentPagerAdapter {
+public class SlidePagerAdapter extends FragmentPagerAdapter {
 
     private List<Fragment> fragmentList;
 
@@ -28,4 +26,5 @@ public class SlidePagerAdapter
     public int getCount() {
         return fragmentList.size();
     }
+
 }
