+++ app/src/main/res/layout-land/page_3.xml
@@ -24,7 +24,7 @@
             android:layout_marginRight="16dp"
             android:layout_marginEnd="16dp"
             android:text="@string/theme_light"
-            android:background="@xml/border" />
+            android:background="@drawable/border" />
 
         <RadioButton
             android:id="@+id/radio_button_dark"
@@ -32,7 +32,7 @@
             android:layout_height="match_parent"
             android:layout_weight="1"
             android:text="@string/theme_dark"
-            android:background="@xml/border" />
+            android:background="@drawable/border" />
 
     </RadioGroup>
