+++ app/src/main/res/layout/activity_welcome_page.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WelcomePage.ActivityWelcomePage">
+    tools:context=".Activities.WelcomePage.ActivityWelcomePage">
 
     <android.support.v4.view.ViewPager
         android:id="@+id/view_pager"
