+++ app/src/main/res/values/colors.xml
@@ -7,4 +7,13 @@
     <color name="activityBackground">#808080</color>
     <color name="white">#FFFFFF</color>
 
+    
+    <color name="primary">#3F51B5</color>
+    <color name="primary_dark">#303F9F</color>
+    <color name="primary_light">#C5CAE9</color>
+    <color name="accent">#FFEB3B</color>
+    <color name="primary_text">#212121</color>
+    <color name="secondary_text">#757575</color>
+    <color name="icons">#FFFFFF</color>
+    <color name="divider">#BDBDBD</color>
 </resources>
